Add files via upload
Counter resetted noch nicht aber sonst fertig, häckchen gesetzt

+++ Timer.c
@@ -10,7 +10,7 @@ void setTimer(uint8_t timer_index,uint32_t time,void (*cb)())
 	timers[timer_index].passedTime = 0;
 }
 
-void startTimer(uint8_t timer_index,uint16_t currentTime)
+void startTimer(uint8_t timer_index,uint32_t currentTime)
 {
 	timers[timer_index].startingTime = currentTime;
 	timers[timer_index].start = 1;
@@ -21,7 +21,7 @@ void cancelTimer(uint8_t timer_index)
 	timers[timer_index].start = 0;
 }
 
-void timerCheck(uint16_t currentTime)
+void timerCheck(uint32_t currentTime)
 {
 
 		for(int i = 0; i<4; i++)
